requirements.txt
appdirs==1.4.4
asgiref==3.2.7
boto3==1.16.16
botocore==1.19.16
certifi==2020.4.5.1
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.0
Django==3.0.6
django-allauth==0.43.0
django-filter==2.4.0
django-import-export==2.4.0
django-storages==1.10.1
djangorestframework==3.12.2
et-xmlfile==1.0.1
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
jdcal==1.4.1
jmespath==0.10.0
MarkupPy==1.14
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.5
Pillow==7.1.2
pipenv==2018.11.26
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
tablib==2.0.0
urllib3==1.25.10
virtualenv==20.0.20
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.6
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0
